+++ src/main/java/visang/dataplatform/dataportal/model/dto/stdcheck/response/StdTermDetailDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class StdTermDetailDto {
-    private Long term_idx;
+    private Integer term_idx;
     private String term_logical_nm;
     private String term_logical_desc;
     private String term_physical_nm;
