+++ src/main/java/visang/dataplatform/dataportal/model/dto/stdcheck/response/StdWordDetailDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class StdWordDetailDto {
-    private Long word_idx;
+    private Integer word_idx;
     private String word_logical_nm;
     private String word_physical_nm;
     private String word_logical_desc;
